python/requirements.txt
pyannote.audio==3.1.1
torch>=2.0.0
numpy>=1.20.0
soundfile>=0.12.1
librosa>=0.10.0
huggingface_hub>=0.16.4 